# api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8090","basePath":"/politikei","tags":[{"name":"home-controller","description":"Home Controller"},{"name":"vote-controller","description":"Vote Controller"},{"name":"proposition-controller","description":"Proposition Controller"},{"name":"basic-error-controller","description":"Basic Error Controller"}],"paths":{"/":{"get":{"tags":["home-controller"],"summary":"home","operationId":"homeUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["home-controller"],"summary":"home","operationId":"homeUsingHEAD","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"post":{"tags":["home-controller"],"summary":"home","operationId":"homeUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["home-controller"],"summary":"home","operationId":"homeUsingPUT","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["home-controller"],"summary":"home","operationId":"homeUsingDELETE","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"options":{"tags":["home-controller"],"summary":"home","operationId":"homeUsingOPTIONS","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"patch":{"tags":["home-controller"],"summary":"home","operationId":"homeUsingPATCH","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/api/proposition":{"get":{"tags":["proposition-controller"],"summary":"getOne","operationId":"getOneUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"details","in":"query","required":false,"type":"object"},{"name":"authenticated","in":"query","required":false,"type":"boolean"},{"name":"propositionId","in":"query","description":"propositionId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Proposition"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["proposition-controller"],"summary":"postNew","operationId":"postNewUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"details","in":"query","required":false,"type":"object"},{"name":"authenticated","in":"query","required":false,"type":"boolean"},{"name":"entityId","in":"query","description":"entityId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Proposition"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["proposition-controller"],"summary":"putUpdate","operationId":"putUpdateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"details","in":"query","required":false,"type":"object"},{"name":"authenticated","in":"query","required":false,"type":"boolean"},{"in":"body","name":"command","description":"command","required":true,"schema":{"$ref":"#/definitions/Proposition"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Proposition"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/proposition/{entityId}":{"get":{"tags":["proposition-controller"],"summary":"getAll","operationId":"getAllUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"details","in":"query","required":false,"type":"object"},{"name":"authenticated","in":"query","required":false,"type":"boolean"},{"name":"entityId","in":"path","description":"entityId","required":true,"type":"string"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":0,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Page«Proposition»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/vote":{"get":{"tags":["vote-controller"],"summary":"getOne","operationId":"getOneUsingGET_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"voteId","in":"query","description":"voteId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Vote"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/vote/down":{"put":{"tags":["vote-controller"],"summary":"voteDown","operationId":"voteDownUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"details","in":"query","required":false,"type":"object"},{"name":"authenticated","in":"query","required":false,"type":"boolean"},{"name":"propositionId","in":"path","description":"propositionId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Vote"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/vote/up":{"put":{"tags":["vote-controller"],"summary":"voteUp","operationId":"voteUpUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"details","in":"query","required":false,"type":"object"},{"name":"authenticated","in":"query","required":false,"type":"boolean"},{"name":"propositionId","in":"path","description":"propositionId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Vote"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/vote/{propositionId}":{"get":{"tags":["vote-controller"],"summary":"getAll","operationId":"getAllUsingGET_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"propositionId","in":"path","description":"propositionId","required":true,"type":"string"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":0,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Page«Vote»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingGET","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingHEAD","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"post":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPOST","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPUT","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingDELETE","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingOPTIONS","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPATCH","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}}},"definitions":{"Vote":{"type":"object","properties":{"id":{"type":"string"},"identityId":{"type":"string"},"propositionId":{"type":"string"},"vote":{"type":"integer","format":"int32"},"voteDate":{"type":"string","format":"date-time"}}},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["100","101","102","103","200","201","202","203","204","205","206","207","208","226","300","301","302","303","304","305","307","308","400","401","402","403","404","405","406","407","408","409","410","411","412","413","414","415","416","417","418","419","420","421","422","423","424","426","428","429","431","451","500","501","502","503","504","505","506","507","508","509","510","511"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}}},"Sort":{"type":"object"},"Page«Vote»":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/definitions/Vote"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"number":{"type":"integer","format":"int32"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"sort":{"$ref":"#/definitions/Sort"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"Page«Proposition»":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/definitions/Proposition"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"number":{"type":"integer","format":"int32"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"sort":{"$ref":"#/definitions/Sort"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"Proposition":{"type":"object","properties":{"docAbstract":{"type":"string"},"docCode":{"type":"string"},"docContent":{"type":"string"},"docType":{"type":"integer","format":"int32"},"entityId":{"type":"string"},"id":{"type":"string"},"votedDown":{"type":"integer","format":"int32"},"votedOther":{"type":"integer","format":"int32"},"votedUp":{"type":"integer","format":"int32"}}},"View":{"type":"object","properties":{"contentType":{"type":"string"}}}}}